requirements.txt
Flask==1.1.2
Flask-Migrate==2.5.3
Flask-Script==2.0.6
Flask-SQLAlchemy==2.4.4
requests==2.24.0
python-dotenv==0.14.0